pyproject.toml
[tool.poetry]
name = "veri-hsl"
version = "0.0.1"
description = "HSL High-frequency positioning example"
authors = ["Lauri <lauri@pm.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
ipykernel = "^6.5.0"
paho-mqtt = "^1.6.1"
psycopg2 = "^2.9.1"
python-dotenv = "^0.19.2"
fastapi = "^0.70.0"
watchtower = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
